remove insights and move tests
