+++ posthog/api/__init__.py
@@ -105,15 +105,14 @@ from ee.clickhouse.views.element import ClickhouseElementViewSet, LegacyClickhou
 from ee.clickhouse.views.events import ClickhouseEventsViewSet, LegacyClickhouseEventsViewSet
 from ee.clickhouse.views.experiments import ClickhouseExperimentsViewSet
 from ee.clickhouse.views.groups import ClickhouseGroupsTypesView, ClickhouseGroupsView
-from ee.clickhouse.views.insights import ClickhouseInsightsViewSet, LegacyClickhouseInsightsViewSet
 from ee.clickhouse.views.paths import ClickhousePathsViewSet, LegacyClickhousePathsViewSet
 from ee.clickhouse.views.person import ClickhousePersonViewSet, LegacyClickhousePersonViewSet
 from ee.clickhouse.views.session_recordings import ClickhouseSessionRecordingViewSet
+from posthog.api.insight import InsightViewSet
 
 # Legacy endpoints CH (to be removed eventually)
 router.register(r"action", LegacyClickhouseActionsViewSet, basename="action")  # Should be completely unused now
 router.register(r"event", LegacyClickhouseEventsViewSet, basename="event")  # Should be completely unused now
-router.register(r"insight", LegacyClickhouseInsightsViewSet, basename="insight")  # Should be completely unused now
 router.register(r"person", LegacyClickhousePersonViewSet, basename="person")
 router.register(r"paths", LegacyClickhousePathsViewSet, basename="paths")
 router.register(r"element", LegacyClickhouseElementViewSet, basename="element")
@@ -123,7 +122,7 @@ projects_router.register(r"actions", ClickhouseActionsViewSet, "project_actions"
 projects_router.register(r"events", ClickhouseEventsViewSet, "project_events", ["team_id"])
 projects_router.register(r"groups", ClickhouseGroupsView, "project_groups", ["team_id"])
 projects_router.register(r"groups_types", ClickhouseGroupsTypesView, "project_groups_types", ["team_id"])
-projects_router.register(r"insights", ClickhouseInsightsViewSet, "project_insights", ["team_id"])
+projects_router.register(r"insights", InsightViewSet, "project_insights", ["team_id"])
 projects_router.register(r"persons", ClickhousePersonViewSet, "project_persons", ["team_id"])
 projects_router.register(r"paths", ClickhousePathsViewSet, "project_paths", ["team_id"])
 projects_router.register(r"elements", ClickhouseElementViewSet, "project_elements", ["team_id"])
